Created userland daemon factory test
